+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "trendradar"
-version = "6.9.1"
+version = "6.10.0"
 description = "TrendRadar - 热点新闻聚合与分析工具"
 requires-python = ">=3.12"
 dependencies = [
