+++ trendradar/__init__.py
@@ -9,5 +9,5 @@ TrendRadar - 热点新闻聚合与分析工具
 
 from trendradar.context import AppContext
 
-__version__ = "6.9.1"
+__version__ = "6.10.0"
 __all__ = ["AppContext", "__version__"]
