+++ trendradar/ai/__init__.py
@@ -7,6 +7,7 @@ TrendRadar AI 模块
 
 from .analyzer import AIAnalyzer, AIAnalysisResult
 from .filter import AIFilter, AIFilterResult
+from .filter_pipeline import AIFilterPipeline
 from .translator import AITranslator, TranslationResult, BatchTranslationResult
 from .formatter import (
     get_ai_analysis_renderer,
@@ -24,6 +25,7 @@ __all__ = [
     # 智能筛选
     "AIFilter",
     "AIFilterResult",
+    "AIFilterPipeline",
     # 翻译器
     "AITranslator",
     "TranslationResult",
