+++ trendradar/ai/formatter.py
@@ -65,19 +65,36 @@ def _format_list_content(text: str) -> str:
     return result
 
 
-def _format_standalone_summaries(summaries: dict) -> str:
-    """格式化独立展示区概括为纯文本行，每个源名称单独一行"""
+def _format_standalone_summaries(
+    summaries: dict, bracket_left: str = "[", bracket_right: str = "]"
+) -> str:
+    """格式化独立展示区概括为纯文本行，每个源名称单独一行
+
+    Args:
+        summaries: 源名称 -> 概括文本 的字典
+        bracket_left/bracket_right: 源名称两侧的方括号字符。飞书卡片 2.0 markdown
+            基于 CommonMark，裸 ``[源名]:`` 会被当作「链接引用定义」整段吞掉，
+            故飞书须传入 HTML 实体 ``&#91;`` ``&#93;``（其余渠道用默认裸方括号）。
+    """
     if not summaries:
         return ""
     lines = []
     for source_name, summary in summaries.items():
         if summary:
-            lines.append(f"[{source_name}]:\n{summary}")
+            lines.append(f"{bracket_left}{source_name}{bracket_right}:\n{summary}")
     return "\n\n".join(lines)
 
 
-def render_ai_analysis_markdown(result: AIAnalysisResult) -> str:
-    """渲染为通用 Markdown 格式（Telegram、企业微信、ntfy、Bark、Slack）"""
+def _render_ai_analysis_markdown_like(
+    result: AIAnalysisResult, standalone_brackets=("[", "]")
+) -> str:
+    """Markdown 系渠道的通用渲染骨架（飞书 / 企业微信 / ntfy / Slack 共用）
+
+    Args:
+        standalone_brackets: 独立源点速览中源名两侧的括号字符。飞书卡片 markdown
+            须用 HTML 实体 ``("&#91;", "&#93;")`` 避免源名被「链接引用定义」吞掉，
+            其余渠道沿用默认裸方括号。
+    """
     if not result.success:
         if result.skipped:
             return f"ℹ️ {result.error}"
@@ -107,49 +124,30 @@ def render_ai_analysis_markdown(result: AIAnalysisResult) -> str:
         )
 
     if result.standalone_summaries:
-        summaries_text = _format_standalone_summaries(result.standalone_summaries)
+        summaries_text = _format_standalone_summaries(
+            result.standalone_summaries, *standalone_brackets
+        )
         if summaries_text:
             lines.extend(["**独立源点速览**", summaries_text])
 
     return "\n".join(lines)
+
+
+def render_ai_analysis_markdown(result: AIAnalysisResult) -> str:
+    """渲染为通用 Markdown 格式（企业微信、ntfy、Slack）"""
+    return _render_ai_analysis_markdown_like(result)
 
 
 def render_ai_analysis_feishu(result: AIAnalysisResult) -> str:
-    """渲染为飞书卡片 Markdown 格式"""
-    if not result.success:
-        if result.skipped:
-            return f"ℹ️ {result.error}"
-        return f"⚠️ AI 分析失败: {result.error}"
+    """渲染为飞书卡片 2.0 markdown 格式
 
-    lines = ["**✨ AI 热点分析**", ""]
-
-    if result.core_trends:
-        lines.extend(["**核心热点态势**", _format_list_content(result.core_trends), ""])
-
-    if result.sentiment_controversy:
-        lines.extend(
-            ["**舆论风向争议**", _format_list_content(result.sentiment_controversy), ""]
-        )
-
-    if result.signals:
-        lines.extend(["**异动与弱信号**", _format_list_content(result.signals), ""])
-
-    if result.rss_insights:
-        lines.extend(
-            ["**RSS 深度洞察**", _format_list_content(result.rss_insights), ""]
-        )
-
-    if result.outlook_strategy:
-        lines.extend(
-            ["**研判策略建议**", _format_list_content(result.outlook_strategy), ""]
-        )
-
-    if result.standalone_summaries:
-        summaries_text = _format_standalone_summaries(result.standalone_summaries)
-        if summaries_text:
-            lines.extend(["**独立源点速览**", summaries_text])
-
-    return "\n".join(lines)
+    飞书卡片 markdown 基于 CommonMark，裸 ``[源名]:`` 会被解析为「链接引用定义」
+    (link reference definition) 而整段不显示，故独立源点速览的源名改用 HTML 实体
+    方括号 ``&#91;`` ``&#93;``（与 report/formatter.py 标题来源标签的处理一致）。
+    """
+    return _render_ai_analysis_markdown_like(
+        result, standalone_brackets=("&#91;", "&#93;")
+    )
 
 
 def render_ai_analysis_dingtalk(result: AIAnalysisResult) -> str:
