+++ version_configs
@@ -1,4 +1,4 @@
-config.yaml=2.3.0
+config.yaml=2.4.0
 timeline.yaml=1.2.0
 frequency_words.txt=1.1.0
 ai_interests.txt=1.1.0
